+++ lib/string_fixer/__init__.py
@@ -21,14 +21,15 @@ def version_lt(a: str, b: str):
 
 
 class QuoteTransformer(cst.CSTTransformer):
-    def __init__(self, target_python: Optional[str] = None):
+    def __init__(self, target_python: Optional[str] = None, prefer_least_escapes = False):
         '''
         Args:
             target_python: which version of python to target. Defaults to current version
         '''
         self._target_python = target_python or f'{sys.version_info.major}.{sys.version_info.minor}'
+        self.prefer_least_escapes = prefer_least_escapes
 
-    def _escape_quote(self, match: re.Match) -> str:
+    def _escape_quote_sub(self, match: re.Match) -> str:
         '''
         Handle quotes, check if they're escaped, escape them if not
         '''
@@ -38,6 +39,35 @@ class QuoteTransformer(cst.CSTTransformer):
             return escapes + quote
         # quote is not escaped. Escape it
         return escapes + (('\\' + quote[0]) * len(quote))
+
+    def _escape_quotes(self, text: str, quote_re: str, quote_len: int):
+        return re.sub(
+            r'(\\*)([%s]{%d,})' % (quote_re, quote_len),
+            self._escape_quote_sub,
+            text,
+            flags=re.MULTILINE,
+        )
+
+    def _unescape_quote_sub(self, match: re.Match) -> str:
+        escapes, quote = match.groups()
+        if not quote:
+            return escapes + quote
+        if len(escapes) % 2 == 0:
+            # quote is unescaped. Do nothing
+            return escapes + quote
+        # quote is escaped. Unescape it
+        return escapes[:-1] + quote
+
+    def _unescape_quotes(self, text: str, quote_re: str, quote_len: int):
+        return re.sub(
+            r'(\\*)([%s]{1,%d})' % (quote_re, quote_len),
+            self._unescape_quote_sub,
+            text,
+            flags=re.MULTILINE,
+        )
+
+    def _count_escapes(self, text: str):
+        return len([i for i in re.findall(r'(\\*)["\']', text) if len(i) % 2 != 0])
 
     def leave_SimpleString(
         self, original_node: cst.SimpleString, updated_node: cst.SimpleString,
@@ -60,15 +90,29 @@ class QuoteTransformer(cst.CSTTransformer):
 
         # remove start and end quotes
         text = updated_node.value[len(updated_node.quote) + len(prefix) : -len(updated_node.quote)]
-
-        text = re.sub(
-            r'(\\*)(["\']{%d,})' % quote_len,
-            self._escape_quote,
-            text,
-            flags=re.MULTILINE,
-        )
-
         new_quote = quote[0] * quote_len
+
+        if 'r' in prefix :
+            if quote in text:
+                # we can't add/remove escapes from rstrings
+                return original_node
+            # if the target quote isn't in the rstring then we can do a simple quote swap
+        else:
+            # unescape as many internal quotes as possible
+            text = orig_text = self._unescape_quotes(text, quote + anti, quote_len)
+            # escape the quotes we want to wrap the string with
+            text = self._escape_quotes(text, quote, quote_len)
+
+            if self.prefer_least_escapes:
+                # escape the current quotes wrapping the string
+                old_style = self._escape_quotes(orig_text, anti, quote_len)
+
+                # if changing quote style would result in more escapes
+                if self._count_escapes(text) > self._count_escapes(old_style):
+                    # set everything back to the old style
+                    text = old_style
+                    new_quote = anti[0] * quote_len
+
         return updated_node.with_changes(value=f'{prefix}{new_quote}{text}{new_quote}')
 
     def leave_FormattedString(
@@ -112,7 +156,7 @@ class QuoteTransformer(cst.CSTTransformer):
             if isinstance(part, cst.FormattedStringText):
                 value = re.sub(
                     r'(\\*)(["\']{%d,})' % 1,
-                    self._escape_quote,
+                    self._escape_quote_sub,
                     part.value,
                     flags=re.MULTILINE,
                 )
@@ -176,9 +220,9 @@ class QuoteTransformer(cst.CSTTransformer):
         return updated_node.with_changes(parts=new_parts, start=f'{original_node.prefix}{quote}', end=quote)
 
 
-def replace_quotes(code: str, target_python: Optional[str] = None) -> str:
+def replace_quotes(code: str, target_python: Optional[str] = None, **kwargs) -> str:
     module = parse_module(code)
-    transformer = QuoteTransformer(target_python)
+    transformer = QuoteTransformer(target_python, **kwargs)
     modified_module = module.visit(transformer)
     return modified_module.code
 
@@ -191,7 +235,7 @@ def process_file(file: Path, config: Config, base_dir: Optional[Path] = None):
     with open(file) as f:
         code = f.read()
 
-    modified = replace_quotes(code, config['target_version'])
+    modified = replace_quotes(code, config['target_version'], prefer_least_escapes=config['prefer_least_escapes'])
 
     if config.get('dry_run', False):
         print('---')