+++ lib/test/snapshots/strings_and_escapes.py
@@ -4,7 +4,7 @@
 'Simple string containing \'another string\''
 'Simple string containing "another string"'
 
-'Simple string containing \"escaped string\"'
+'Simple string containing "escaped string"'
 'Simple string containing \'escaped string\''
 
 
@@ -12,3 +12,5 @@
 '''Unescaped \\"quotation\\" in string'''
 
 '''\'\'\'Nested triple quote\'\'\''''
+
+'String containing \'one string\' and "another style" of string and \'a third\''